pyproject.toml
[tool.poetry]
name = "pyar"
version = "0.2.0"
description = "Real-time OpenCV-based Augmented Reality demo using ArUco markers."
authors = ["thejavasaurus@gmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2"
opencv-contrib-python = "^4.12"
imutils = "^0.5.4"
Pillow = "^11.3"
reportlab = "^4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
